Update type info with using typed apis
Switch to np.float64 as default datatype

+++ src/quadrotor_mpc/acados_wrapper/utils.py
@@ -77,6 +77,6 @@ def run_simulation(
             model.model_update(u_setpoint, simulation_dt)
 
     tout = np.asarray(tout)
-    yout = {k: np.asarray(v) for k, v in yout.items()}
+    yout = {k: np.asarray(v, dtype=np.float64) for k, v in yout.items()}
 
     return tout, yout
